+++ src/utils/promptUtils.ts
@@ -72,3 +72,21 @@ export async function promptValues(structureVariables: Variable[], structureOpti
 
     return { variables, optionals };
 }
+
+export async function promptShowSkippedItems(skippedItems: { [key: string]: string }, createdItemsCount: number) {
+    if (skippedItems) {
+        const skippedItemsCount = Object.keys(skippedItems).length;
+
+        const showSkippedItems = await vscode.window.showInformationMessage(`Successfully created ${createdItemsCount} item${createdItemsCount !== 1 ? 's' : ''} and skipped ${skippedItemsCount} item${skippedItemsCount !== 1 ? 's' : ''}.`, 'Show overview', 'OK') === 'Show overview';
+
+        if (showSkippedItems) {
+            let skippedItemsOverview: string = `Skipped item${skippedItemsCount !== 1 ? 's' : ''}:`;
+            for (const item in skippedItems) {
+                skippedItemsOverview += `\n\n- ${item}:\n      ${skippedItems[item]}`;
+            }
+            await vscode.window.showInformationMessage(skippedItemsOverview, { modal: true });
+        }
+    } else {
+        vscode.window.showInformationMessage(`Successfully created ${createdItemsCount} item${createdItemsCount !== 1 ? 's' : ''}.`);
+    }
+}